Update 37-sudoku-solver.rs
FIRST WORKING VERSION

It can solve the leetcode example input! 

I haven't actually verified that the output matches the expected output, but it runs to completion and all of the internal consistency and correctness assertions I've added leave me highly confident.

It's not _finished_ -- I am 100% positive that there are Sudoku puzzles that it cannot solve -- but I'm definitely on the right track.

The next thing I want to tackle is cleaning up the "look for exactly one solution left" case.  That code was written much earlier, when the information tracked by the solver's iteration was much less detailed.

After that, I need to implement logic for indirect elimination tactics.  For example, within any region (i.e. within any row, column, or box):

- For any set S(U) of N unsolved symbols within a region,
- There is a set S(C) of M candidate cells where those symbols may be placed.
- If M=N, then those M cells must contain _exactly_ those N symbols, and nothing else.
- So we can rule out any candidate symbol in S(M) that isn't part of S(U)

For example, if you have a row with cells like: [5, 7, 8] [5, 7, 9] and no other cell in that row can have a 5 or a 7, then we can rule out that 8 from the first column and that 9 from the second column.

This is useful for any N that is less than the total number of unsolved symbols in the region.